Adding support for Autonomous Linux

+++ common/bash/prereqs.sh
@@ -17,6 +17,16 @@ __is_cloud_shell(){
     fi
 }
 
+__is_autonomous_linux(){
+    sudo alx show
+    if [ $? -eq 0 ]; then
+        echo "Autonomous Linux detected"
+        return 0
+    else
+        return 1
+    fi
+}
+
 __is_ubuntu_or_debian() {
   if [ -f /etc/os-release ]; then
     . /etc/os-release
@@ -96,6 +106,12 @@ __is_oci_cli_installed(){
         else
             return 1
         fi
+    elif __is_autonomous_linux; then            
+        if [ -d /usr/bin/oci ]; then
+            return 0
+        else
+            return 1
+        fi
     else
         echo 'Error: This is not a supported system.' >&2
         exit 1
@@ -118,17 +134,19 @@ __source_bashrc
 if ! __is_oci_cli_installed ; then    
   echo '============== CLI Pre-Requisites =============='
     if __is_redhat_or_centos ; then
-        if ! __is_cloud_shell ; then      
-            echo 'Installing OCI CLI in RedHat or CentOS...'
-            sudo runuser -l opc -c 'mkdir -p /home/opc/oci_cli'
-            sudo runuser -l opc -c 'wget https://raw.githubusercontent.com/oracle/oci-cli/master/scripts/install/install.sh'
-            sudo runuser -l opc -c 'chmod +x install.sh'
-            sudo runuser -l opc -c '/home/opc/install.sh --install-dir /home/opc/oci_cli/lib/oracle-cli --exec-dir /home/opc/oci_cli/bin --accept-all-defaults'
-            sudo runuser -l opc -c 'cp -rl /home/opc/bin /home/opc/oci_cli'
-            sudo runuser -l opc -c 'rm -r /home/opc/bin'
-            sudo runuser -l opc -c 'mkdir -p /home/opc/.oci'
-            sudo runuser -l opc -c 'touch /home/opc/.oci/config'
-            sudo runuser -l opc -c 'oci setup repair-file-permissions --file /home/opc/.oci/config'
+        if ! __is_cloud_shell ; then  
+            if ! __is_autonomous_linux; then    
+                echo 'Installing OCI CLI in RedHat or CentOS...'
+                sudo runuser -l opc -c 'mkdir -p /home/opc/oci_cli'
+                sudo runuser -l opc -c 'wget https://raw.githubusercontent.com/oracle/oci-cli/master/scripts/install/install.sh'
+                sudo runuser -l opc -c 'chmod +x install.sh'
+                sudo runuser -l opc -c '/home/opc/install.sh --install-dir /home/opc/oci_cli/lib/oracle-cli --exec-dir /home/opc/oci_cli/bin --accept-all-defaults'
+                sudo runuser -l opc -c 'cp -rl /home/opc/bin /home/opc/oci_cli'
+                sudo runuser -l opc -c 'rm -r /home/opc/bin'
+                sudo runuser -l opc -c 'mkdir -p /home/opc/.oci'
+                sudo runuser -l opc -c 'touch /home/opc/.oci/config'
+                sudo runuser -l opc -c 'oci setup repair-file-permissions --file /home/opc/.oci/config'
+            fi
         fi
     elif __is_ubuntu_or_debian ; then
             echo 'Installing OCI CLI in Ubuntu or Debian...'
